ingestors/requirements.txt
# HEGO Ingestors — Python dependencies
elasticsearch[async]>=8.0.0,<9.0.0
requests>=2.31.0,<3.0.0
pycti>=6.0.0,<7.0.0
python-dotenv>=1.0.0,<2.0.0
tenacity>=8.2.0,<9.0.0
schedule>=1.2.0,<2.0.0
python-dateutil>=2.8.0,<3.0.0
